Add TOOL_HOSTS constant for per-tool DNS mapping and integrate into MitmToolCard component
